+++ ProjectSQ/Models/Processor.cs
@@ -7,7 +7,7 @@ namespace ProjectSQ.Models
         #region prepare processor
         public static int startStack = 60000, endStack = 65535, currentStackPointer = 60000;
 
-        public static readonly Dictionary<string, int> registerDictionary = new Dictionary<string, int> {
+        public static readonly Dictionary<string, short> registerDictionary = new Dictionary<string, short> {
             {"reg1" , 0 }, {"reg2" , 0 }, {"reg3" , 0 },{"reg4" , 0 }, {"reg5" , 0 }, {"reg6" , 0 }, {"reg7" , 0 }, {"reg8" , 0 }
         };
         public static bool Equal { get; set; }
